Update README.md
The following is a text description of the system you are being tasked to develop. The system requirements – explicit and implicit – are included this document, and they need to be identified and reflected in your Enhanced Entity-Relationship Diagram (EERD).

You are being asked to design and develop a system to monitor deliveries of grocery products to customers. This system will support a “third party” grocery service. Customers will place orders with the service. The service will coordinate with grocery stores to find the products (at variable – but hopefully the lowest – prices) and arrange for a drone to deliver the products to the customer. On delivery, the store will be paid electronically by the customer.

Your system will support various types of users, including customers and employees of the grocery stores. All users must be either customers and/or employees – our system will not keep track of any other types of user accounts. All users will have distinct usernames when accessing the system.

Attributes that are used to identify entities in our system will normally consist of forty (40) or fewer alphanumeric characters in some regular pattern/format. This will be the default format for entity-identifying and “entity unique” attributes in our system unless otherwise noted.

We will also maintain the first name, last name, address, and birthdate of each user. The birthdate will be represented in the “yyyy-mm-dd” date format. This will be the default format for dates in our system unless otherwise noted. We will use the birthdate to help authenticate users, and send birthday greeting messages along with discounts. Some users might not want to share their birthdate, but we must have their name and address for drone deliveries and/or tax and salary purposes